1.91.8 (2018.2.0f2)
Build-Id: df3cdeaf-4a4d-4d0d-99fb-ee16c2948252
Build-Sha256: c1ebbded1e9fda6bd616072019dcfd311d6bd055891e03afebc0026a60dc5422

+++ Assembly-CSharp/BodyAccessoriesController.cs
@@ -355,7 +355,7 @@ public class BodyAccessoriesController
 			foreach (KeyValuePair<object, object> accessoryDatum in AccessoryData)
 			{
 				Dictionary<object, object> dictionary = (Dictionary<object, object>)accessoryDatum.Value;
-				if (accessorySlot == (AccessorySlotType)(int)dictionary[AvatarAccessoryData.Slot.ToString("d")])
+				if (accessorySlot == (AccessorySlotType)dictionary[AvatarAccessoryData.Slot.ToString("d")])
 				{
 					accessoryStreamingAssetsId = (int)dictionary[AvatarAccessoryData.InventoryID.ToString("d")];
 				}
