+++ Assembly-CSharp/MVJetPack.cs
@@ -232,7 +232,7 @@ public class MVJetPack : MVVehicleBase
 
 		private void OverheatUpdate(bool thrust)
 		{
-			JetModeType jetModeType = (JetModeType)(byte)owner.JetMode.Value;
+			JetModeType jetModeType = (JetModeType)owner.JetMode.Value;
 			JetModeType jetModeType2 = JetModeType.Off;
 			if (thrust)
 			{
@@ -468,7 +468,7 @@ public class MVJetPack : MVVehicleBase
 		{
 			Debug.LogError("WoData does not contain jetPackType ");
 		}
-		return (JetPackType)(byte)dictionary["jetPackType"];
+		return (JetPackType)dictionary["jetPackType"];
 	}
 
 	private static VehicleBaseObject GetPickupPrefabName(Dictionary<object, object> data)
