+++ Assembly-CSharp/TeamRequirement.cs
@@ -58,7 +58,7 @@ public class TeamRequirement : UseRequirement
 	{
 		if (data.ContainsKey("team"))
 		{
-			team = (MVTeam)(int)data["team"];
+			team = (MVTeam)data["team"];
 			if (team == MVTeam.None)
 			{
 				Dictionary<object, object> dictionary = new Dictionary<object, object>();
